Did a variety of cleanup.
Angular unit testing has been setup. Only one test has been written. So there's a bunch
of work to do create unit tests. The description language is jasmine and it's provided by
the jasmine-rails gem. To run a test, enter the comment

RAILS_ENV=test rake spec:javascript

Additional items such a node.js might need setup.

Add a player model to indicate when a user is a player in a game.

Don't think ActiveRecord:Associations are correct for the game model. I don't think the Game
really has_many players. But rather each Game belongs to one or more players. This needs to be fixed.

Started to add in some tradition rails testing with rspect. We're going to need this also. This would help uncover the problem above.